// ionic-app/src/app/models/interfaces.ts
export interface IRecord<T> {
	id: string;
	data?: T;
}
// ...
export type MetricColor = 'primary' | 'secondary' | 'tertiary' | 'success' | 'danger' | 'warning';

export interface IBoolMetricVal {
	label: string;
	color: MetricColor;
}

export interface IBoolMetric {
	true: IBoolMetricVal;
	false: IBoolMetricVal;
}

export interface ITeamMetric {
	id?: string;
	title: string;
	colorTrue?: MetricColor;
	colorFalse?: MetricColor;
	type: 'bool' | 'int' | 'options';
	mode: 'personal' | 'team';
	bool?: IBoolMetric;
	min?: number;
	max?: number;
}

export interface IMeetingInfo {
	id: string;
	stage: string;
	started?: string;
	finished?: string;
}

export interface ITeamMeetings {
	scrum?: IMemberInfo;
	retrospective?: IMeetingInfo;
}

export interface ITeam {
	title: string;
	userIds: string[];
	memberIds?: string[];
	members: IMemberInfo[];
	metrics: ITeamMetric[];
	active?: ITeamMeetings;
	last?: ITeamMeetings;
	upcomingRetro?: { itemsByUserAndType?: { [user: string]: { [itemType: string]: number } } };
}

export interface IAvatar {
	gravatar?: string;
	external?: {
		provider?: string;
		url?: string;
	};
}

export interface IMemberInfo {
	id: string;
	role?: MemberRole;
	roles: MemberRole[];
	email?: string;
	uid?: string;
	title: string;
	avatar?: IAvatar;
}

export interface IMember {
	title: string;
	roles: MemberRole[];
	email?: string;
	userId?: string;
	avatar?: IAvatar;
}

export type TaskType = 'done' | 'risk' | 'todo' | 'plan' | 'qna' | 'kudos';

interface Modified {
	by: string;
	on: string;
}

export interface ITask {
	id: string;
	title: string;
	by?: IBy;
	progress?: number;
	added?: Modified;
	thumbUps?: string[];
	lastModified?: Modified;
	comments?: ITaskComment[];
}

interface IQuestion extends ITask {
	isAnswered?: boolean;
	isAnswerAccepted?: boolean;
}

export interface IBy {
	memberId?: string;
	userId?: string;
	title: string;
}

export interface ITaskComment {
	id: string;
	by: IBy;
	message: string;
}

export interface IScrumStatusMember {
	id: string;
	roles?: MemberRole[];
	title: string;
	avatar?: IAvatar;
}

export interface IStatus {
	member: IScrumStatusMember;
	byType: {
		plan?: ITask[]; // For now for UI only
		done?: ITask[];
		todo?: ITask[];
		risk?: ITask[];
		qna?: IQuestion[];
		kudos?: ITask[];
	}; // TODO: consider mapped object type
}


export interface IByUser {
	uid: string
	title?: string;
}

export type IMeetingMember = IMemberInfo;

export interface IMeeting {
	timer?: ITimerState;
	readonly userIds: string[];
	readonly members?: IMeetingMember[];
}

export interface IScrum extends IMeeting {  // Key as: YYYY-MM-DD
	scrumIds?: {
		prev?: string; // 'YYYY-MM-DD'
		next?: string; // 'YYYY-MM-DD'
	};
	readonly risksCount?: number;
	readonly questionsCount?: number;
	statuses: IStatus[];
}

export interface ITimerState {
	status?: TimerStatus;
	elapsedSeconds?: number;
	activeMemberId?: string;
	at?: Timestamp | string;
	by?: IByUser;
	secondsByMember?: { [member: string]: number };
	isToggling?: boolean;
}